# docs/swagger/stasogg-1.0.0-swagger.yaml
swagger: '2.0'
info:
  description: API componente stasogg
  version: 1.0.0
  title: STASOGG
  contact:
    email: andrea.fontanelli@eng.it
# tags are used for organizing operations
tags:
- name: test
  description: test resources
- name: proprietario
  description: soggetto proprietario di un veicolo
- name: soggetto
  description: soggetto
  

# basePath: /api
# schemes:
#    - http

paths:
  
  /test/resources:
    get:
      tags:
        - test
      summary: Test resources
      description: Effettua il test dei servizi richiamati
      operationId: testResources
      produces:
      - application/json
      responses:
        200:
          schema:
            $ref: '#/definitions/GenericResponse'
          description: OK
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'
  
  /soggetto/{codiceFiscale}:
    get:
      tags:
        - soggetto
      summary: ricerca Soggetto
      description: Recupera il soggetto dato il codice fiscale
      operationId: ricercaSoggetto
      produces:
      - application/json
      parameters:
        - in: path
          name: codiceFiscale
          type: string
          required: true
          description: "Codice fiscale proprietario"
      responses:
        200:
          description: Elenco soggetti trovati
          schema:
            type: array
            items:
              $ref: '#/definitions/Proprietario'
        400:
          description: "Parametro non valorizzato"
        404:
          description: "Soggetto non trovato"
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'    
  /soggetto/{idGms}/{idProprietario}:
    get:
      tags:
        - soggetto
      summary: ricerca Soggetto dato idGMS e id Proprietario
      description: Recupera il soggetto dato l'id GMS e id Proprietario
      operationId: loadById
      produces:
      - application/json
      parameters:
        - in: path
          name: idGms
          type: integer
          format: int64
          required: true
          description: "ID GMS"
        - in: path
          name: idProprietario
          type: integer
          format: int64
          required: true
          description: "ID PROPRIETARIO"
      responses:
        200:
          description: Elenco soggetti trovati
          schema:
            type: array
            items:
              $ref: '#/definitions/Proprietario'
        400:
          description: "Parametro non valorizzato"
        404:
          description: "Soggetto non trovato"
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'   
  
  /soggetto/pg:
    get:
      tags:
        - soggetto
      summary: ricerca Soggetto Persona giuridica
      description: Recupera il soggetto impresa data la denominazione
      operationId: ricercaSoggettoPG
      produces:
      - application/json
      parameters:
        - in: query
          name: denominazione
          type: string
          required: true
          description: "La denominazione dell'impresa"
      responses:
        200:
          description: Elenco soggetti trovati
          schema:
            type: array
            items:
              $ref: '#/definitions/Proprietario'
        400:
          description: "Parametro non valorizzato"
        404:
          description: "Soggetto non trovato"
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorDto' 
            
  /soggetto/pf:
    get:
      tags:
        - soggetto
      summary: ricerca Soggetto Persona Fisica
      description: Recupera il soggetto persona fisica dato nome e cognome
      operationId: ricercaSoggettoPF
      produces:
      - application/json
      parameters:
        - in: query
          name: nome
          type: string
          required: true
          description: "Il nome del soggetto"
        - in: query
          name: cognome
          type: string
          required: true
          description: "Il cognome del soggetto"
      responses:
        200:
          description: Elenco soggetti trovati
          schema:
            type: array
            items:
              $ref: '#/definitions/Proprietario'
        400:
          description: "Parametro non valorizzato"
        404:
          description: "Soggetto non trovato"
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'    
          
definitions:
  Proprietario:
    type: object
    description: Entità che definisce il soggetto proprietario di un veicolo
    properties:
      id:
        type: integer
        format: int64
        description: pk proprietario
  GenericResponse:
    type: object
    description: Risposta di un servizio generico
    properties:
      message:
        type: string
        description: Contiene il messaggio eventuale del servizio chiamato
        example: 'Servizio ha risposto correttamente'
  ErrorDto:
    type: object
    description: Rappresenta l'errore restituito da servizio
    properties:
      status:
        type: integer
        description: status http restituito
        example: '500, 400, ecc'
      code:
        type: string
        description: code http restituito
        example: 'Internal server error, ecc'
      description:
        type: string
        description: la descrizione dell'errore
        example: 'Internal server error, ecc'
      errorDetails:
        type: array
        description: i dettagli errore
        items:
          $ref: '#/definitions/ErrorDetail'
  ErrorDetail:
    type: object
    description: rappresenta il dettaglio errore restituito in caso di errori di validazione sui campi di input
    properties:
      field:
        type: string
        description: il nome del campo per cui è falita la validazione
        example: 'name, surname'
      error:
        type: string
        description: l'errore di validazione effettivo
        example: 'campo obbligatorio, ecc'
      description:
        type: string
        description: la descrizione dell'errore
        example: 'Internal server error, ecc'
  
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
# basePath: /studiofontanelli/stacomunicazionesrv/1.0.0
schemes:
 - https
 - http
# Added by API Auto Mocking Plugin
# basePath: /andreafontanelli/stacomunicazionesrv/1.0.0
# Added by API Auto Mocking Plugin
# basePath: /andreafontanelli/staconsultazionesrv/1.0.0
# Added by API Auto Mocking Plugin
basePath: /andreafontanelli/stasogg/1.0.0